레시피 상세 조회 구현 1
- 레시피 조회 테스트 green
- 레시피 조회 실패 exception, errorMessage 정의
- 레시피 관련 실패 exception handle 추가
- 레시피 조회 실패 테스트 green
- 레시피 추가 presentation layer 수정

+++ src/main/java/com/my/bob/core/domain/recipe/service/RecipeService.java
@@ -10,5 +10,6 @@ public interface RecipeService {
 
     Page<RecipeListItemDto> getRecipes(Pageable pageable, RecipeSearchDto dto);
 
-    RecipeDto getRecipe(Long id);
+    RecipeDto getRecipe(Integer id);
+
 }
